unified callback parameter definition of size_t and uint32, now uses size_t
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@18869 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/os/drivers/USB2.h
@@ -74,7 +74,7 @@ struct usb_configuration_info {
 };
 
 typedef void (*usb_callback_func)(void *cookie, uint32 status, void *data,
-	uint32 actualLength);
+	size_t actualLength);
 
 
 /* The usb_module_info represents the public API of the USB Stack. */
